0.1.0-alpha9 adds lookup-key to data sources
This is intended to support tables where you may need to lookup a
primary key prior to update! / insert! based on other values in the
row (such as one-to-many / many-to-many tables where it is not generally
known when the update is scheduled, whether or not the entry exists or
not, and needs to be replaced or added).